requirements.txt
fastapi==0.111.0
pydantic==2.7.1
pymongo==4.7.2
python-dotenv==1.0.1
uvicorn==0.15.0